+++ src/main/scala/cromwell/engine/backend/sge/SgeBackendCall.scala
@@ -6,6 +6,7 @@ import cromwell.engine.{AbortRegistrationFunction, WorkflowDescriptor}
 import wdl4s.CallInputs
 
 import scala.concurrent.{ExecutionContext, Future}
+import scala.util.Try
 
 case class SgeBackendCall(backend: SgeBackend,
                           workflowDescriptor: WorkflowDescriptor,
@@ -19,6 +20,11 @@ case class SgeBackendCall(backend: SgeBackend,
   val returnCode = callRootPath.resolve("rc")
   val engineFunctions: SgeEngineFunctions = new SgeEngineFunctions(callRootPath, stdout, stderr, workflowDescriptor.ioManager)
   callRootPath.toFile.mkdirs
+
+  def instantiateCommand: Try[String] = {
+    val backendInputs = backend.adjustInputPaths(this)
+    call.instantiateCommandLine(backendInputs, engineFunctions)
+  }
 
   override def execute(implicit ec: ExecutionContext) = backend.execute(this)